+++ src/components/pages/ServerManager.vue
@@ -18,10 +18,6 @@
             </el-table-column>
             <el-table-column label="主机名" prop="serverName" align="left"
                         min-width="200">
-                <!-- <template slot-scope="scope">
-                    <router-link :to="`/vm/virtual/${scope.row.id}`"
-                                :class="$style.link">{{scope.row.vmName}}</router-link>
-                </template> -->
             </el-table-column>
             <el-table-column prop="hostName" align="left"
                             label="所属设备"
@@ -143,21 +139,15 @@ export default {
             })
         },
         addBackupPlan(data) {
-            console.log(data);
             let plan = Object.assign({}, data);
             let vmIds = this.currentSelect.map( e => {
                 return e.id;
             });
             plan.vmList = vmIds;
-            console.log(plan);
             this.btnLoading = true;
             createMultipleVirtualBackupPlan(plan)
                 .then(res => {
-                const { message } = res.data;
-                // // 刷新情况下可能会出现两个添加后的计划
-                // if (this.backupPlans.findIndex(plan => plan.id === backupPlan.id) === -1) {
-                //     this.backupPlans.unshift(backupPlan)
-                // }
+                    const { message } = res.data;
                     this.backupPlanCreateModalVisible = false;
                     this.$confirm(
                         '备份计划添加成功，是否跳转到备份计划页面',
@@ -173,11 +163,11 @@ export default {
                     });
                 })
                 .catch(error => {
-                this.$message.error(error);
-                return false;
+                    this.$message.error(error);
+                    return false;
                 })
                 .then(() => {
-                this.btnLoading = false;
+                    this.btnLoading = false;
                 });
         },
         submitServerFn(data){
@@ -185,19 +175,50 @@ export default {
             this.btnLoading = true;
             addServer(server).then( res => {
                 const {data, message} = res.data;
-                if(this.serverTableData.findIndex( e => e.id === data.id) === -1){
-                    let fData = data
-                    fData.disabled = false;
-                    fData.icon = 'el-icon-refresh';
-                    this.serverTableData.unshift(fData);
-                    let d = Object.assign([], this.serverTableData);
-                    this.serverTableData = [];
-                    this.$nextTick( () => {
-                        this.serverTableData = d;
-                    });
-                }
+                // if(this.serverTableData.findIndex( e => e.id === data.id) === -1){
+                //     let fData = data
+                //     fData.disabled = false;
+                //     fData.icon = 'el-icon-refresh';
+                //     this.serverTableData.unshift(fData);
+                //     let d = Object.assign([], this.serverTableData);
+                //     this.serverTableData = [];
+                //     this.$nextTick( () => {
+                //         this.serverTableData = d;
+                //     });
+                // }
+                this.btnLoading = false;
                 this.serverModalVisible = false;
-                this.$message.success(message);
+                this.$confirm(
+                    '主机添加成功，请确认是否需要扫描虚拟机？',
+                    '提示',
+                    {
+                        confirmButtonText: '确定',
+                        cancelButtonText: '取消',
+                        type: 'success',
+                        beforeClose: (action, instance, done) => {
+                            if (action === 'confirm') {
+                                instance.confirmButtonLoading = true;
+                                instance.confirmButtonText = '扫描中...';
+                                rescan(data).then( resp => {
+                                    this.$message.success( resp.data.message);
+                                    // this.fetchData();
+                                }).catch( error => {
+                                    this.$message.error( error)
+                                }).then( () => {
+                                    done();
+                                    instance.confirmButtonLoading = false;
+                                })
+                            }else {
+                                done();
+                            }
+                        }
+                    }
+                )
+                .then(action => {
+                    this.fetchData();
+                }).catch( () => {
+                    this.fetchData();
+                });
             }).catch( error => {
                 this.$message.error(error);
             }).then( () => {
@@ -241,12 +262,22 @@ export default {
             })
         },
         deleteServer(scope){
-            deleteServer(scope.row.id).then( res => {
-                this.$message.success( res.data.message);
-                this.fetchData();
-            }).catch( error => {
-                this.$message.error( error);
-            })
+            this.$confirm(
+                '请确认是否删除',
+                '提示',
+                {
+                    confirmButtonText: '确定',
+                    cancelButtonText: '取消',
+                    type: 'warning'
+                }
+            ).then(() => {
+                deleteServer(scope.row.id).then( res => {
+                    this.$message.success( res.data.message);
+                    this.fetchData();
+                }).catch( error => {
+                    this.$message.error( error);
+                });
+            }).catch( () => {});
         }
     }
 }
